test(unit_tests): split mixed files and update documentation
Split 9 mixed test files (containing both unit and integration tests) into
separate unit_tests/unit/ and unit_tests/integration/ counterparts:

- test_azure_vm_provider_reboot.py (3 unit, 1 integration)
- test_cluster.py (69 unit, 14 integration)
- test_config.py (61 unit, 24 integration)
- test_dedicated_hosts.py (6 unit, 1 integration)
- test_events.py (33 unit, 1 integration)
- test_find_ami_equivalent.py (6 unit, 4 integration)
- test_latte_thread.py (54 unit, 4 integration)
- test_scylla_bench_thread.py (18 unit, 3 integration)
- test_version_utils.py (314 unit, 12 integration)

Update AGENTS.md, writing-unit-tests and writing-integration-tests skills,
and plan doc to reflect the new directory structure.

1961 tests collected, 0 errors.
