+++ unit_tests/integration/test_dedicated_hosts.py
@@ -1,0 +1,21 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2024 ScyllaDB
+
+import pytest
+
+from sdcm.provision.aws.dedicated_host import SCTDedicatedHosts
+
+
+@pytest.mark.integration
+def test_release_by_tag_integration():
+    SCTDedicatedHosts.release_by_tags({"TestId": "no-existing"}, regions=["us-east-1", "eu-west-1"], dry_run=True)
